requirements.txt
PyGithub
semgrep
openai
python-dotenv
